feat: use FancyPageRoute globally
