+++ lib/app/app.dart
@@ -3,6 +3,7 @@ export 'data.dart';
 export 'datetime_utils.dart';
 export 'hive.dart';
 export 'logger.dart';
+export 'routing.dart' show FancyRoute, FancyRouteBuilder;
 export 'services/api_network.dart';
 export 'services/network.dart';
 export 'services/snack_bar.dart';
@@ -18,7 +19,7 @@ export 'widgets/empty_state.dart';
 export 'widgets/error_widgets.dart';
 export 'widgets/fade_in.dart';
 export 'widgets/form.dart';
-export 'widgets/page_route.dart' show TopLevelPageRoute;
 export 'widgets/scaffold.dart';
 export 'widgets/schulcloud_app.dart' show SchulCloudApp, SignedInScreen;
 export 'widgets/text.dart';
+export 'widgets/top_level_page_route.dart';
