+++ lib/app/widgets/page_route.dart
@@ -1,15 +1,52 @@
+// This is a stripped-down implementation of [CupertinoPageRoute] that allows
+// for swiping back anywhere on the page.
+
+import 'dart:math';
+import 'dart:ui' show lerpDouble;
+
 import 'package:flutter/material.dart';
+import 'package:flutter/foundation.dart';
+import 'package:flutter/gestures.dart';
+import 'package:flutter/animation.dart' show Curves;
 
-import 'top_level_screen_wrapper.dart';
+const double _kMinFlingVelocity = 1; // Screen widths per second.
 
-class TopLevelPageRoute<T> extends PageRoute<T> {
-  TopLevelPageRoute({
+// An eyeballed value for the maximum time it takes for a page to animate forward
+// if the user releases a page mid swipe.
+const int _kMaxDroppedSwipePageForwardAnimationTime = 800; // Milliseconds.
+
+// The maximum time for a page to get reset to it's original position if the
+// user releases a page mid swipe.
+const int _kMaxPageBackAnimationTime = 300; // Milliseconds.
+
+// Offset from offscreen to the right to fully on screen.
+final Animatable<Offset> _kRightMiddleTween = Tween<Offset>(
+  begin: Offset(1, 0),
+  end: Offset.zero,
+);
+
+// Offset from fully on screen to 1/3 offscreen to the left.
+final Animatable<Offset> _kMiddleLeftTween = Tween<Offset>(
+  begin: Offset.zero,
+  end: Offset(-1.0 / 3, 0),
+);
+
+class FancyPageRoute<T> extends PageRoute<T> {
+  FancyPageRoute({
     @required this.builder,
-    @required RouteSettings settings,
+    RouteSettings settings,
   })  : assert(builder != null),
-        super(settings: settings);
+        assert(opaque),
+        super(settings: settings, fullscreenDialog: false);
 
+  /// Builds the primary contents of the route.
   final WidgetBuilder builder;
+
+  @override
+  bool get maintainState => true;
+
+  @override
+  Duration get transitionDuration => Duration(milliseconds: 400);
 
   @override
   Color get barrierColor => null;
@@ -18,30 +55,351 @@ class TopLevelPageRoute<T> extends PageRoute<T> {
   String get barrierLabel => null;
 
   @override
+  bool canTransitionFrom(TransitionRoute<dynamic> previousRoute) =>
+      previousRoute is FancyPageRoute;
+
+  @override
+  bool canTransitionTo(TransitionRoute<dynamic> nextRoute) =>
+      nextRoute is FancyPageRoute;
+
+  static bool isPopGestureInProgress(PageRoute<dynamic> route) =>
+      route.navigator.userGestureInProgress;
+
+  bool get popGestureInProgress => isPopGestureInProgress(this);
+
+  bool get popGestureEnabled => _isPopGestureEnabled(this);
+
+  static bool _isPopGestureEnabled<T>(PageRoute<T> route) {
+    // If there's nothing to go back to, then obviously we don't support the
+    // back gesture.
+    if (route.isFirst) {
+      return false;
+    }
+
+    // If we're in an animation already, we cannot be manually swiped.
+    if (route.animation.status != AnimationStatus.completed) {
+      return false;
+    }
+
+    // If we're being popped into, we also cannot be swiped until the pop above
+    // it completes. This translates to our secondary animation being
+    // dismissed.
+    if (route.secondaryAnimation.status != AnimationStatus.dismissed) {
+      return false;
+    }
+
+    // If we're in a gesture already, we cannot start another.
+    if (isPopGestureInProgress(route)) {
+      return false;
+    }
+
+    // Looks like a back gesture would be welcome!
+    return true;
+  }
+
+  @override
   Widget buildPage(
+      BuildContext context, Animation<double> _, Animation<double> __) {
+    return Semantics(
+      scopesRoute: true,
+      explicitChildNodes: true,
+      child: builder(context),
+    );
+  }
+
+  // Called by _FancyBackGestureDetector when a pop ("back") drag start
+  // gesture is detected. The returned controller handles all of the subsequent
+  // drag events.
+  static _FancyBackGestureController<T> _startPopGesture<T>(
+      PageRoute<T> route) {
+    assert(_isPopGestureEnabled(route));
+
+    return _FancyBackGestureController<T>(
+      navigator: route.navigator,
+      controller: route.controller, // protected access
+    );
+  }
+
+  static Widget buildPageTransitions<T>(
+    PageRoute<T> route,
     BuildContext context,
     Animation<double> animation,
     Animation<double> secondaryAnimation,
-  ) {
-    return TopLevelScreenWrapper(child: builder(context));
-  }
+    Widget child,
+  ) =>
+      FancyPageTransition(
+        primaryRouteAnimation: animation,
+        secondaryRouteAnimation: secondaryAnimation,
+        // Check if the route has an animation that's currently participating
+        // in a back swipe gesture.
+        // In the middle of a back gesture drag, let the transition be linear to
+        // match finger motions.
+        linearTransition: isPopGestureInProgress(route),
+        child: _FancyBackGestureDetector<T>(
+          enabledCallback: () => _isPopGestureEnabled<T>(route),
+          onStartPopGesture: () => _startPopGesture<T>(route),
+          child: child,
+        ),
+      );
 
   @override
   Widget buildTransitions(BuildContext context, Animation<double> animation,
       Animation<double> secondaryAnimation, Widget child) {
-    final anim = CurvedAnimation(parent: animation, curve: Curves.easeOutQuad);
-    return FadeTransition(
-      opacity: anim,
-      child: ScaleTransition(
-        scale: Tween<double>(begin: 0.9, end: 1).animate(anim),
-        child: child,
+    return buildPageTransitions<T>(
+        this, context, animation, secondaryAnimation, child);
+  }
+}
+
+class FancyPageTransition extends StatelessWidget {
+  FancyPageTransition({
+    Key key,
+    @required Animation<double> primaryRouteAnimation,
+    @required Animation<double> secondaryRouteAnimation,
+    @required this.child,
+    @required bool linearTransition,
+  })  : assert(linearTransition != null),
+        _primaryPositionAnimation = (linearTransition
+                ? primaryRouteAnimation
+                : CurvedAnimation(
+                    parent: primaryRouteAnimation,
+                    curve: Curves.linearToEaseOut,
+                    reverseCurve: Curves.easeInToLinear,
+                  ))
+            .drive(_kRightMiddleTween),
+        _secondaryPositionAnimation = (linearTransition
+                ? secondaryRouteAnimation
+                : CurvedAnimation(
+                    parent: secondaryRouteAnimation,
+                    curve: Curves.linearToEaseOut,
+                    reverseCurve: Curves.easeInToLinear,
+                  ))
+            .drive(_kMiddleLeftTween),
+        super(key: key);
+
+  // When this page is coming in to cover another page.
+  final Animation<Offset> _primaryPositionAnimation;
+  // When this page is becoming covered by another page.
+  final Animation<Offset> _secondaryPositionAnimation;
+
+  /// The widget below this widget in the tree.
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) {
+    assert(debugCheckHasDirectionality(context));
+    final textDirection = Directionality.of(context);
+    return SlideTransition(
+      position: _secondaryPositionAnimation,
+      textDirection: textDirection,
+      transformHitTests: false,
+      child: SlideTransition(
+        position: _primaryPositionAnimation,
+        textDirection: textDirection,
+        child: Material(elevation: 12, child: child),
       ),
     );
   }
+}
+
+/// This is the widget side of [_FancyBackGestureController].
+///
+/// This widget provides a gesture recognizer which, when it determines the
+/// route can be closed with a back gesture, creates the controller and
+/// feeds it the input from the gesture recognizer.
+///
+/// The gesture data is converted from absolute coordinates to logical
+/// coordinates by this widget.
+///
+/// The type `T` specifies the return type of the route with which this gesture
+/// detector is associated.
+class _FancyBackGestureDetector<T> extends StatefulWidget {
+  const _FancyBackGestureDetector({
+    Key key,
+    @required this.enabledCallback,
+    @required this.onStartPopGesture,
+    @required this.child,
+  })  : assert(enabledCallback != null),
+        assert(onStartPopGesture != null),
+        assert(child != null),
+        super(key: key);
+
+  final Widget child;
+
+  final ValueGetter<bool> enabledCallback;
+
+  final ValueGetter<_FancyBackGestureController<T>> onStartPopGesture;
 
   @override
-  bool get maintainState => true;
+  _FancyBackGestureDetectorState<T> createState() =>
+      _FancyBackGestureDetectorState<T>();
+}
+
+class _FancyBackGestureDetectorState<T>
+    extends State<_FancyBackGestureDetector<T>> {
+  _FancyBackGestureController<T> _backGestureController;
+
+  HorizontalDragGestureRecognizer _recognizer;
 
   @override
-  Duration get transitionDuration => Duration(milliseconds: 300);
+  void initState() {
+    super.initState();
+    _recognizer = HorizontalDragGestureRecognizer(debugOwner: this)
+      ..onStart = _handleDragStart
+      ..onUpdate = _handleDragUpdate
+      ..onEnd = _handleDragEnd
+      ..onCancel = _handleDragCancel;
+  }
+
+  @override
+  void dispose() {
+    _recognizer.dispose();
+    super.dispose();
+  }
+
+  void _handleDragStart(DragStartDetails details) {
+    assert(mounted);
+    assert(_backGestureController == null);
+    _backGestureController = widget.onStartPopGesture();
+  }
+
+  void _handleDragUpdate(DragUpdateDetails details) {
+    assert(mounted);
+    assert(_backGestureController != null);
+    _backGestureController.dragUpdate(
+        _convertToLogical(details.primaryDelta / context.size.width));
+  }
+
+  void _handleDragEnd(DragEndDetails details) {
+    assert(mounted);
+    assert(_backGestureController != null);
+    _backGestureController.dragEnd(_convertToLogical(
+        details.velocity.pixelsPerSecond.dx / context.size.width));
+    _backGestureController = null;
+  }
+
+  void _handleDragCancel() {
+    assert(mounted);
+    // This can be called even if start is not called, paired with the "down" event
+    // that we don't consider here.
+    _backGestureController?.dragEnd(0);
+    _backGestureController = null;
+  }
+
+  void _handlePointerDown(PointerDownEvent event) {
+    if (widget.enabledCallback()) {
+      _recognizer.addPointer(event);
+    }
+  }
+
+  double _convertToLogical(double value) {
+    switch (Directionality.of(context)) {
+      case TextDirection.rtl:
+        return -value;
+      case TextDirection.ltr:
+        return value;
+      default:
+        throw StateError('No Directionality ancestor above $this.');
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    assert(debugCheckHasDirectionality(context));
+    return Stack(
+      fit: StackFit.passthrough,
+      children: <Widget>[
+        widget.child,
+        Positioned.fill(
+          child: Listener(
+            onPointerDown: _handlePointerDown,
+            behavior: HitTestBehavior.translucent,
+          ),
+        ),
+      ],
+    );
+  }
+}
+
+class _FancyBackGestureController<T> {
+  _FancyBackGestureController({
+    @required this.navigator,
+    @required this.controller,
+  })  : assert(navigator != null),
+        assert(controller != null) {
+    navigator.didStartUserGesture();
+  }
+
+  final AnimationController controller;
+  final NavigatorState navigator;
+
+  /// The drag gesture has changed by [delta]. The total range of the drag
+  /// should be 0.0 to 1.0.
+  void dragUpdate(double delta) {
+    controller.value = (controller.value - delta).clamp(0.0001, 0.9999);
+  }
+
+  /// The drag gesture has ended with a horizontal motion of [velocity] as a
+  /// fraction of screen width per second.
+  void dragEnd(double velocity) {
+    // Fling in the appropriate direction.
+    // AnimationController.fling is guaranteed to take at least one frame.
+    const Curve animationCurve = Curves.fastLinearToSlowEaseIn;
+    bool animateForward;
+
+    // If the user releases the page before mid screen with sufficient velocity,
+    // or after mid screen, we should animate the page out. Otherwise, the page
+    // should be animated back in.
+    if (velocity.abs() >= _kMinFlingVelocity) {
+      animateForward = velocity <= 0;
+    } else {
+      animateForward = controller.value > 0.5;
+    }
+
+    if (animateForward) {
+      // The closer the panel is to dismissing, the shorter the animation is.
+      // We want to cap the animation time, but we want to use a linear curve
+      // to determine it.
+      final droppedPageForwardAnimationTime = min(
+        lerpDouble(
+                _kMaxDroppedSwipePageForwardAnimationTime, 0, controller.value)
+            .floor(),
+        _kMaxPageBackAnimationTime,
+      );
+      controller.animateTo(
+        1,
+        duration: Duration(milliseconds: droppedPageForwardAnimationTime),
+        curve: animationCurve,
+      );
+    } else {
+      // This route is destined to pop at this point. Reuse navigator's pop.
+      navigator.pop();
+
+      // The popping may have finished inline if already at the target destination.
+      if (controller.isAnimating) {
+        // Otherwise, use a custom popping animation duration and curve.
+        final droppedPageBackAnimationTime = lerpDouble(
+                0, _kMaxDroppedSwipePageForwardAnimationTime, controller.value)
+            .floor();
+        controller.animateBack(
+          0,
+          duration: Duration(milliseconds: droppedPageBackAnimationTime),
+          curve: animationCurve,
+        );
+      }
+    }
+
+    if (controller.isAnimating) {
+      // Keep the userGestureInProgress in true state so we don't change the
+      // curve of the page transition mid-flight since FancyPageTransition
+      // depends on userGestureInProgress.
+      AnimationStatusListener animationStatusCallback;
+      animationStatusCallback = (status) {
+        navigator.didStopUserGesture();
+        controller.removeStatusListener(animationStatusCallback);
+      };
+      controller.addStatusListener(animationStatusCallback);
+    } else {
+      navigator.didStopUserGesture();
+    }
+  }
 }
