+++ lib/file/file.dart
@@ -5,4 +5,3 @@ export 'widgets/choose_destination.dart';
 export 'widgets/file_browser.dart';
 export 'widgets/file_tile.dart';
 export 'widgets/files_screen.dart';
-export 'widgets/page_route.dart';
