+++ lib/file/routes.dart
@@ -5,46 +5,36 @@ import 'package:schulcloud/course/course.dart';
 import 'data.dart';
 import 'widgets/file_browser.dart';
 import 'widgets/files_screen.dart';
-import 'widgets/page_route.dart';
 
 Route _buildSubdirRoute(Id<Entity> Function(RouteResult result) ownerGetter) {
-  return Route(
+  return FancyRoute(
     matcher: Matcher.path('{parentId}'),
-    builder: (result) => FileBrowserPageRoute(
-      builder: (_) => FileBrowser(FilePath(
-        ownerGetter(result),
-        Id<File>(result['parentId']),
-      )),
-      settings: result.settings,
-    ),
+    builder: (_, result) => FileBrowser(FilePath(
+      ownerGetter(result),
+      Id<File>(result['parentId']),
+    )),
   );
 }
 
-final fileRoutes = Route(
+final fileRoutes = FancyRoute(
   matcher: Matcher.path('files'),
-  materialBuilder: (_, __) => FilesScreen(),
+  builder: (_, __) => FilesScreen(),
   routes: [
-    Route(
+    FancyRoute(
       matcher: Matcher.path('my'),
-      builder: (result) => FileBrowserPageRoute(
-        builder: (_) => FileBrowser(FilePath(services.storage.userId)),
-        settings: result.settings,
-      ),
+      builder: (_, result) => FileBrowser(FilePath(services.storage.userId)),
       routes: [
         _buildSubdirRoute((_) => services.storage.userId),
       ],
     ),
-    Route(
+    FancyRoute(
       matcher: Matcher.path('courses'),
-      materialBuilder: (_, __) => FilesScreen(),
+      builder: (_, __) => FilesScreen(),
       routes: [
-        Route(
+        FancyRoute(
           matcher: Matcher.path('{courseId}'),
-          builder: (result) => FileBrowserPageRoute(
-            builder: (_) =>
-                FileBrowser(FilePath(Id<Course>(result['courseId']))),
-            settings: result.settings,
-          ),
+          builder: (_, result) =>
+              FileBrowser(FilePath(Id<Course>(result['courseId']))),
           routes: [
             _buildSubdirRoute((result) => Id<Course>(result['courseId'])),
           ],
