+++ lib/settings/routes.dart
@@ -1,8 +1,9 @@
 import 'package:flutter_deep_linking/flutter_deep_linking.dart';
+import 'package:schulcloud/app/app.dart';
 
 import 'widgets/settings_screen.dart';
 
-final settingsRoutes = Route(
+final settingsRoutes = FancyRoute(
   matcher: Matcher.path('settings'),
-  materialBuilder: (_, __) => SettingsScreen(),
+  builder: (_, __) => SettingsScreen(),
 );
